+++ samples/drivers/flash_shell/src/main.c
@@ -9,7 +9,7 @@
 
 #include <zephyr.h>
 #include <misc/printk.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <flash.h>
 #include <device.h>
 #include <soc.h>
