+++ subsys/bluetooth/shell/flash.c
@@ -10,7 +10,7 @@
  */
 
 #include <zephyr.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <misc/printk.h>
 
 #include <stdlib.h>
