+++ subsys/bluetooth/shell/ticker.c
@@ -10,7 +10,7 @@
  */
 
 #include <zephyr.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <misc/printk.h>
 
 #if defined(CONFIG_SOC_FAMILY_NRF)
