+++ subsys/fs/shell.c
@@ -9,7 +9,7 @@
 #include <string.h>
 
 #include <misc/printk.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <init.h>
 #include <fs.h>
 #include <stdio.h>
