+++ subsys/shell/shell_service.c
@@ -13,7 +13,7 @@
  */
 
 #include <misc/printk.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <init.h>
 
 #define SHELL_PROMPT "shell> "
